+++ addOns/quickstart/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Changed
+- Default modern spider browser to Firefox headless.
 
+### Fixed
+- Persist the default browsers by ID instead of by (translatable) name, so the setting is restored correctly when ZAP is not running in English.
 
 ## [59] - 2026-07-13
 ### Added